+++ styles.js
@@ -1,11 +1,13 @@
 export * as bootstrap from './palette/bootstrap.js';
 export * as gnome from './palette/gnome.js';
 export * as palette from './palette/aegis.js';
-export * from './button.js';
 export { reset } from './reset.js';
-export { sheetToFile, sheetToLink, registerProperties } from './utils.js';
-export * from './consts.js';
-export * from './misc.js';
-export * from './forms.js';
 export * from './properties.js';
 export * from  './theme.js';
+export * from './button.js';
+export * from './misc.js';
+export * from './forms.js';
+export * from './scrollbar.js';
+export * from './animations.js';
+export * from './consts.js';
+export { sheetToFile, sheetToLink, registerProperties } from './utils.js';